third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-log-doesnt-run-microtasks.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function testFunction()
{
    Promise.resolve().then(function(){ console.log(239); });
    console.log(42);
    console.log(43);
}

function test()
{
    InspectorTest.sendCommandOrDie("Runtime.enable", {});
    InspectorTest.eventHandler["Runtime.consoleAPICalled"] = messageAdded;
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { "expression": "testFunction()" });
    InspectorTest.sendCommandOrDie("Runtime.evaluate", { "expression": "setTimeout(() => console.log(\"finished\"), 0)" });

    function messageAdded(result)
    {
        InspectorTest.logObject(result.params.args[0]);
        if (result.params.args[0].value === "finished")
            InspectorTest.completeTest();
    }
}
</script>
</head>
<body onLoad="runTest();">
Check that console.log doesn't run microtasks.
</body>
</html>
